requirements.txt
fastapi
uvicorn
line-bot-sdk
beautifulsoup4
requests